+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `aiopg_listen.connect_func` to simplify initialization [#5](https://github.com/Pliner/aiopg-listen/pull/5)
 * Rename consumer to listener [#6](https://github.com/Pliner/aiopg-listen/pull/6)
+* Do not crash if handler fails [#7](https://github.com/Pliner/aiopg-listen/pull/7)
 
 ## v0.0.1 (2021-07-25)