+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.0.3 (2021-08-10)
+
+* Allow suppressing timeout by aiopg_listen.NO_TIMEOUT [#9](https://github.com/Pliner/aiopg-listen/pull/9)
+
 ## v0.0.2 (2021-07-25)
 
 * Add `aiopg_listen.connect_func` to simplify initialization [#5](https://github.com/Pliner/aiopg-listen/pull/5)